RenderEngine no longer ignores sensor pose in camera body (#19398)
GeometryState didn't test its Render*Image responsibilities. As it turns
out, they were defective. Specifically, the camera pose should be a
concatenation of *three* transforms.

Given frames for the camera's parent (P), body (B), and sensor (S). The
pose of the camera in the world should be X_WP * X_PB * X_BS. It was
omitting the last transform. Generally, the sensor offset from the body
has been identity so we've been "getting away with it".

- Incidentally enables back face culling in RenderEngineGl.
  - If the camera was inside a mesh (because, for example, the pose was
    not what one was expected), the view would be full of black because
    one would see the un-illuminated back of the mesh.
  - Enabling this is good; it'll help eliminate z-fighting for "skinny"
    geometry and no one should have been relying on it.
  - It might also improve performance as more triangles get dismissed for
    pointing the wrong way.