Make RenderEngineGl intelligently threadsafe
When a RenderEngineGl is cloned, instead of *sharing* an instance of
OpenGlContext with its clones, each engine clone has a unique clone of the
original engine's OpenGlContext. The cloned contexts are unique in that
they can be bound to different threads, but they share the same underlying
OpenGl objects, stored on the GPU.

Because RenderEngineGl no longer shares the exact same OpenGl context with
the original, some additional work needs to be done to the clone to make
sure it is fully prepared:

  - It's OpenGl state needs to be configured (back face culling, etc.)
  - Vertex array objects need to be recreated from the vertex buffers
    shared in the OpenGl context -- by design, vertex array objects don't
    get shared.
    - Additional infrastructure has been added so the recreation of array
      objects from buffers is possible.

This means, the private copy constructor for RenderEngineGl is, itself,
insufficient to create a working copy. Only the public Clone() method
should be used.

A new test has been added to test the thread safety invariants of both
OpenGlContext and RenderEngineGl.

TODO:
 - Basic render engine unit tests are failing.