Make RenderEngineGl copy constructor complete
RenderEngineGl has a private copy constructor that can do *most* of the
work to copy itself, but not all. There are various OpenGl maintenance
tasks that have to be done in conjunction with copying C++ constructs.

This adopts a strategy of exploiting C++'s policy for class member
initialization during construction and copying so that the RenderEngineGl
knows little of the necessary work and that it happens vis a defaulted
copy constructor.

It introduces various types to manage the OpenGlContext (binding and
unbinding it as appropriate) and internal classes which can do the
OpenGl maintenance as part of the copying.

To that end, we move all of the management of OpenGlGeometry into a new
GeometryLibrary that patches the geometry when copied.

TODO:
 - unit tests on GeometryLibrary.


We want to automate the copying logic of RenderEngineGl to minimize the
amount of code in RenderEngineGl for duplicating itself. Ideally, we want
the copy constructor to produce a completely viable new instance. This:

 - Introduce a context wrapper machinery.
   - A wrapper that automatically binds the OpenGlContext.
   - A class that automatically unbinds the OpenGlContext.
   - The wrapper is the first member, the unbinder is the last member.
   - So, initializaiton and copying of all members inbetween are guaranteed
     a bound OpenGL context.
  - Introduce a member responsible for configuring the OpenGL state (e.g.,
    back face culling, etc.)

TODO:
 - Geometry needs to go into its own GeometryLibrary so that it can
   implement its own copy constructor to update its vertex array objects.

WIP Geometry library

Moving the geometry into a geometry library. This now builds.

TODO:
 - Unit tests
 - Devise plan for getting from where we are today to use this.
   - possibly factor out the copy constructor madness for follow up.