Rename SolutionResult::kUnknownError to kSolverSpecificError
I had a student trip on this; it was incredibly frustrating to be
working on a mathematical program only to have the solver apparently
return "kUnknownError".  What they needed, instead, was a hint that
they should look at the solver details, etc, to figure out the error.

Reformat docstrings so that the actually show up in pydrake.

Add deprecation marker (and test) for C++.

Restore pydrake binding with deprecation documentation.