Fix caching for autodiff in direct collocation
My previous attempt at caching failed to address the fact that the
autodiff gradients would be different when the status/inputs were
passed in different places inside the variable list for the
constraint. Now we implement the chain rule locally so that the cache
is kept warm with gradients only with respect to the current
variables.

Resolves #19396.