Implement SapHolonomicConstraint in terms of new APIs introduced in #19402
The new implementation fixes a bug in the functional form of the cost, making it consistent with its analytical gradient and Hessian.
This bug would only appear in holononimic constraints with impulse limits, such as joint stiction and PD control with limits.
Given we haven't made those features available in master yet, no one using public APIs (rather than internal:: source) was affected by it.
Still, we decided to label this PR as a +fix to keep a record of it.