Implement MeshcatPoseSliders
This provides a Systems interface for using sliders in meshcat to
create a pose on the output port. This has been used extensively in my
manipulation course notes (e.g. for end-effector teleop in meshcat),
and will soon power an upcoming tutorial in Drake. This is closely
related to the existing `multibody::meshcat::JointSliders`
implementation.

Resolves #17784.