Start refactoring push_release script
Create a new push_release script written in Python. This takes over
mirroring the .tar and .deb artifacts (and their respective hashes) to
S3; the latter in particular replaces what was formerly a manual process
due to the names being different. This also removes local specification
of the artifact set and local recalculation of the hashes; both now use
whatever is associated with the GitHub release.

Also, remove the actions handled by the new script from the old script.
Eventually we expect to move the docker and apt steps to this script as
well.