Implements multibody::RigidTransformSelector
The MultibodyPlant body_poses output port provides a vector of all
body poses, but often a downstream system just wants to take a single
pose on the input port. This system provides that capability.

While RigidTransform lives in drake::math, I felt that this system
should live in drake::multibody since it is likely to have
SpatialVector and SpatialAcceleration equivalents alongside it soon.