OpenGlInstance now references OpenGlGeometry (#19508)
Previously, each OpenGlInstance contained a *copy* of the OpenGlGeometry
that it instantiated. While the cost of a few ints isn't overly
burdensome, this will be an obstacle in future efforts to clone the
OpenGl context.

RenderEngineGl now maintains a vector of OpenGlGeometry and each instance
merely gets an index into that vector. These indices are amenable to
copying without requiring changing the references.