MultibodyPlant no longer assigns perception color (#19533)
Geometry with perception roles, if they have no definition of color,
should defer to the `RenderEngine`'s logic and default values to pick
up a color. MultibodyPlant shouldn't be playing a role in that regard.

In this case, MultibodyPlant is assuming that if the illustration
properties have a color defined, the perception role should pick up the
same. This is correct so far as our parsing cannot yet distinguish between
illustration and perception roles (they're all just "visual").