Disallow partial Dirichlet BC for an FEM node
Instead, we require the full state (3 DoFs) of a vertex is specified
when a vertex is put under a Dirichlet boundary condition.

The flexibility of putting individual DoFs of a vertex under BC turns
out to be rather useless in practice as our public APIs only allow users
to describe BCs at the level of vertices.

This also paves the way for transitioning to the new data structure of
storing FEM tangent matrix.