[inverse_kinematics] Locked joints are immobile
Locked joints are held fixed during IK (using the their current value
from the context).

Locked joints trump position limits; it's OK to lock outside of the
limit, and the program will still solve.

When a quaternion is locked, we lock its decision variables to the
normalized value and do not add a norm constraint.