MeshcatVisualizer sets visibility directly without events
Previously, the visibility parameter was not respected (a) before the
initialization event or (b) if delete_on_initialization_event was
false.

fixup!

Update visibility exactly when we first transmit shapes to Meshcat.
(Don't add to the scene tree panel too early.)

On the same tack, don't update visibility during re-initialization.
We don't want the dummy note in the scene tree panel then, either.