Implements MultibodyPlant::DeclareBodyPoseOutputPort
This allows a user of MultibodyPlant to declare a new output port that
contains just a single body pose, instead of using the vector of poses
from the existing body_poses output port.  This design was discussed
in #19520.

As a consequence of this (and future features which make system ports
/ diagrams more maleable, we need to add a check in the
LeafOutputPort::DoEval to ensure that the cache entry exists. Without
this check, the user could easily run into a segfault if they create a
context before declaring the output port.