push_release: get GPG fingerprint from public key (#19630)
Modify the logic in push_release that expected the GPG key fingerprint
to be manually entered. While preserving the security of the private key
is good, the identifying fingerprint is part of the public key, and
therefore there is no reason to not simply fetch the public key in order
to obtain the fingerprint, thus removing a manual step.