VPolytope and IrisInConfigurationSpace support Mesh geometry.
Previously they supported Convex geometry; now we extend that support
to Mesh.

For VPolytope: The semantics of this class are already clear that the
set is defined by the convex hull of the points. We even had a test
case of passing a non-convex mesh as Convex to prove it. This PR
simply adds the natural support.

For IrisInConfigurationSpace: Since IRIS is attempting to certify
non-collision, taking the convex hull of the mesh is on the correct
side of conservatism. Having this support will improve usability for
folks bringing their own geometry to IRIS.

Related to #19514.