Add functionality to merge one glTF file into another (#19642)
Given two nlohmann::json representations of a glTF file, this modifies one
of the json trees to include the elements of the second glTF file. glTF
files use internal references (via indices) to relate different elements.
The indicies of the added glTF file all get offset to account for the
elements in the first glTF file.