Adds some text logging to IrisInConfigurationSpace (#19673)
For long-running calls to IRIS, this gives some much needed
feedback. It also can give some insights on if/where IRIS is getting
bogged down.

This PR also clarifies the usage of the term "seed" vs "samples" vs
"searches", which got even more confusing when we start reporting
excessive numbers of searches.

The code does not change the computation (and therefore does not
change the tests).