Rewrite push_release logic for docker (#19631)
Rewrite logic to push the release docker tags in Python, and to rely on
the staging tags we are newly creating. This hopefully makes the logic
easier to follow, and allows for a dry-run. More importantly, it
eliminates the need to associate a release with a nightly build, which
we really shouldn't have been doing in the first place, and simplifies
things for the release manager.

Also remove an outdated instruction from the README, and change the
aptly snapshot to use the source version rather than the corresponding
nightly date for the release's snapshot name.