Add initial documentation about geometric file support in Drake
This provides a new doxygen module for discussing the *geometric* file
formats that Drake supports. The ultimate goal is to allow the user to
be informed as possible on how to get the best possible outcome.

This is not a complete draft. Nor is it final where details are provided.