Implements MakeSemidefiniteRelaxation (#19632)
Given a (typically nonconvex) MathematicalProgram, this method
constructs a new MathematicalProgram which implements the SDP
relaxation of the original program.

This initial implementation supports the case of quadratic objectives
and constraints; it's an extremely important (and quite general) case.

Also touches up a few of the matching lines in the Spectrahedron class
with feedback from code review for this method.