Fix IrisInConfigurationSpace hit and run. (#19728)
When using configuration obstacles, it was possible to add a face to
the polyhedron and move the guess outside of the region. This caused
the hit and run sampler to fail.  Adds a test to cover it.

This PR also touches up the guess logic in all three loops
(configuration obstacles, task obstacles, and additional constraints)
to be consistent.

This PR also fixes the broken StartingEllipse test (it depended on a
false assumption about the invariance across different iterations).