Compute MD ordering with priority nodes
Compute the minimum degree ordering of a sparsity pattern under the
constraint that certain nodes are eliminated before others. This is
useful for computing Schur complement as a side effect of a
right-looking Cholesky factorization as the variables w.r.t which the
Schur complement is computed must be eliminated first.