[geometry] Fix MeshcatVisualizer initialization of alpha slider
When the visualized geometry all has 0% alpha, we need to bump it up
to 100% alpha immediately after the geometry is first added, so that
the slider value special case for all-zero is in effect even before
the user drags the slider at all.

This should also fix the gedanken-bug where the alpha slider is <100%
and scene graph serial number evolves; in that case, we need to sync
again so the demotion is in effect without making the user drags the
slider again.