Meshcat can specify the environment map
Environment maps allow for a higher fidelity of the visualized geometry,
particularly useful for objects defined with physically-based rendering
(PBR) material properties.

This gives the Meshcat class the ability to set and clear such an
environment variable (with the attendant bindings). Furthermore,
model_visualizer and meldis have been extended to take an optional
environment map as a parameter (implicitly testing the bindings).

The actual implementation is tested in meshcat_manual_test.cc where an
environment gets loaded and then subsequently dismissed.