CameraConfig takes RenderEngineParams
Previously, the nature of the render engine was given by a single string:
naming the render engine class to use. It makes the RenderEngine impossible
to configure.

This swaps the renderer_class for render_params: render engine parameters
are contained as a variant. With this, the render engine can be fully
configured in the yaml file.

There are some attendant changes:

1. CameraConfig
  - the fields background and renderer_class are now ignored and should
    be removed. TODO: simply cut? Deprecate some way?
2. The various RenderEngine*Params handle `default_label` differently.
   - RenderLabel was, itself, not serializable.
   - Serializing it (for this purpose) had *no* value.
     - Only two render lable values are valid as parameters for a render
       engine. And even if the appropriate value is known, it can't be
       initialized (easily).
   - In the parameters, we replace the RenderLabel with a string which
     can contain either "unspecified" or "don't_care".
   - RenderEngine maps the strings to actual RenderLabel values.