CameraConfig can use RenderEngineParams
Previously, the nature of the render engine was given by a single string:
naming the render engine class to use. It makes the RenderEngine impossible
to configure.

This extends the `renderer_class` field to also take instances of the
various RenderEngine*Params structs. Now the engines can be fully specified
in yaml.

This also extends the set of renderer classes to include GltfCLient.

Updated the documentation on CameraConfig::background to clarify its
scope in the new world.