Implements planning::VisibilityGraph()
This constructs a visibility graph by running
CollisionChecker::CheckEdgeCollsionFree to create the adjacency
matrix, with support for cpu parallelization.

In order to facilitate using this method (with parallelization) from
Python, we add an optional distance_function_weights to the
CollisionCheckerParams. This allows python callers to construct a
weighted Euclidean distance function that can be used in parallel
executions.