[geometry] Deprecate RenderEngineGltfClient's default_label option (#19906)
Unfortunately, it's not possible to mark it with DRAKE_DEPRECATED
due to warnings from the implicitly-generated default constructor.

Switch the default default_label to kDontCare. The only two permitted
choices are kDontCare or kUnspecified; the RenderLabel docs disavow
kUnspecified, and the Drake RPC already prefers kDontCare.

Drop default_label from Serialize(), which means that the params are
compatible now with YAML serialization.