[workspace] Build vendored clp_internal from source
Deprecate the clp external.

Having discovered during Clp that we don't actually have <ieeefp.h>
available on any of our platforms, fix that same mistake in Ipopt
build rules. It's currently dead code there, but would be a time bomb
on upgrades in case it ever did start to be used.