Pipes constraints forces through MbP and fixes the computation of reaction forces.
It makes use of the infrastructure at the SAP solver level to report constraint forces within SapDriver so that we can compute the aggregate multibody forces due to constraints.
With this information, we can reliably compute reaction forces including all terms (force elements and constraints, including contact) in the discrete update.
This PR also resolves #13888 making the computation of reaction forces maintainable. New constraints will immediately be included in reaction forces.

With this PR we also include discrete hydro forces along with every other solver constraint in the computation of reaction forces, effectively resolving #19068.

In summary:
 - Closes #19068 (discrete hydroelastic forces are included in reaction forces)
 - Closes #18749 (Include hydroelastic in the computation of reaction forces)
 - Closes #19435 (SAP Infrastructure to report multibody forces)
 - Closes #13888 (Fix MbP's CalcReactionForce() and make it maintainable)