RenderEngineVtk consumes and renders glTF (#19961)
A glTF file can contain an arbitrary number of meshes. Those meshes are
lumped together in a single assembly and mapped to the GeometryId. As the
pose of the geometry changes, it is the pose of that assembly that changes.

Two fundamental changes:

1. GeometryId no longer maps to vtkActor and now maps to vtkProp3D.
  - Sometimes the vtkProp3D pointer points to vtkActor and sometimes
    to vtkAssembly.
  - It can't *always* be vtkAssembly as a superset, because VTK does not
    export vtkAssembly to glTF correctly.
    Bug report: discourse.vtk.org/t/bug-in-vtkgltfexporter/12052
  - This requires some extra work in copying because of VTK's weird
    "copy" semantics. It doesn't have a Clone() method which creates a
    new object from an existing; it has a ShallowCopy() method which
    writes one existing object into another -- you must already know the
    concrete type of the copy target. Blech.
  - The protected actors() method has become props(), reflecting the
    change.
2. We're using the `vtkGLTFImporter`.

This also simplified the RenderEngine cloning logic using ShallowCopy()
from VTK.

Incidentally, this is a minimum first step in supporting .obj files with
multiple objects/materials. When we parse objs, we can return multiple
meshes with single materials and map them to parts in assembly like we
have done for glTFs.