Add lights to RenderEngineVtkParams (#19948)
This allows lights to be configured in RenderEngineVtk like they are in
RenderEngineGl.

Calling the `SetDefaultLightPosition()` now spews a warning and does
little else. render_benchmark.cc was the only code in Drake that
exercised this API. It is not part of RenderEngine's public API, just
part of its protected API. This is a precursor to removing it.

The light tests are basically copied from RenderEngineGl's tests (with
small tweaks to make them applicable to VTK).