geometry/render_gltf_client opts into clang format (#20009)
Respell the test to get around clang-format quirks:

The function type with an r-value parameter sometimes confuses clang
format. When included in a function parameter list, it's perfectly happy.

When declaring a variable of that type (as in):

  std::function<void(json*, json&&)> merge = MergeMeshes;

clang format puts a space between "json" and "&&"

  std::function<void(json*, json &&)> merge = MergeMeshes;

Apparently, in this context, it doesn't recognize json as a type name
so it feels the && is the boolean operator and not the r-value ref.
In other contexts is does just fine. By defining the alias, we get around
the erratic treatment.

Clang-format 17 allows us to enumerate non-obvious type names. Although,
it's not clear that that's an unfettered win.
https://clang.llvm.org/docs/ClangFormatStyleOptions.html#typenames