[setup] Change clang-format to a testonly dependency
This should allow users to rebuild from source with fewer headaches.

On macOS, it removes 1 of 2 remaining build-time dependencies on the
custom RobotLocomotion homebrew tap. We anticipate dropping the other
item (VTK) from the custom tap soon, as well.

On Ubuntu, it avoids a hard-coded clang-format version. This should be
more portable for users who build on a non-Ubuntu distribution.

Developers who want to run Drake's full test suite will still need
this dependency; this is only a change for *users* who build from
source.

With this change, the build-time vector_gen can no longer autoformat
its generated code. With a few minor edits to the tool, the output
still looks passable without the autoformat.