[vtk] Unify include spelling, add comments, fix deps (#20024)
VTK splits its code into libraries, but doesn't use subdir include
paths to reflect the which library has which header. This makes it
difficult to properly update BUILD files as the include statements
evolve. For clarity, we now mark all VTK includes with the library
they come from. We can also add some missing deps now as a result.

How should a future developer know which comment (deps) to use for a
given VTK header?

For a VTK header we already use elsewhere in Drake, the easy answer is
to cargo cult from the existing Drake include statement when including
the same VTK header in a new Drake file.

If a developer wants to use a particular VTK header for the first time
inside Drake, the trick is to find the header in the VTK source tree
(e.g., https://github.com/Kitware/VTK/blob/master/IO/Image/vtkImageExport.h)
and so `#include <vtkImageExport.h>` gets the comment `// vtkIOImage`
based on the parent directory names.