[sensors] Deprecate symbolic expression images (#20023)
Their only effect is to clutter up our switch-case statements with an
extra condition that always throws an exception.

Eigen should be a sufficient alternative for users who actually need a
two-dimensional array of expressions. The only advantages of Image<>
over Array<> are semantic labeling of the pixels (color, depth, label)
and multi-channel pixels (rgb) -- neither of which apply here anyway.