PR [workspace] Add vtk_internal build from source
This commit does NOT switch Drake to use the VTK build from source.

Instead, we'll only compile vtkImageIO and then validate it with a
temporary unit test. Future work will switch the various parts of
Drake to use the VTK source build.

The goal here is to get the build system infrastructure under control,
so that it's easy to tweak and apply it down the road.