Fix RenderEngineVtk upside down glTF textures
It has been shown that vtkAssembly leads to textures being rendered
upside down. See
https://discourse.vtk.org/t/vtkgltfimporter-loads-textures-upside-down/12113
The fix we're applying is to get in front of of that and flip the
textures ourselves when importing the glTF.

This PR further modifies rainbow_box.gltf so that its texture is no
longer vertically symmetric (therefore the GltfSupport test now tests
whether the texture is flipped or not).

It also provides a novel glTF file that contains all texture types and
some instructions for doing a visual check on whether the flipping
code works as advertised.