DeformableDriver uses the new FemSolver
Also cleans up the FemSolver public interface.

FemSolver stores the scratch data and the solution as data members to
ensure compatibility of with the FEM model being solved for.

FemSolver::AdvanceOneTimeStep now only takes the previous FEM
state and the set of non-participating vertices as input and FemSolver
offers `next_fem_state` and `next_schur_complement()` functions to
inspect the solution.