Refactor FemSolver to clean up its public interface
FemSolver stores the scratch data and the solution as data members to
ensure compatibility of with the FEM model being solved for.
As a result, FemSolver::AdvanceOneTimeStep only takes the previous FEM
state and the set of non-participating vertices as input and FemSolver
offers a `next_state_and_schur_complement()` function to inspect the
solution.