MeshcatVisualizer and Meldis rely on meshcat APIs to handle alpha sliders
Previously, meshcat visualizer was trying to modulate the opacity of
geometries by storing the underlying color/opacity and subsequently setting
it by setting the "color" property. This works for primtive shapes but
breaks down for complex geometries (e.g., obj with mtl files and glTF
files).

Instead, MeshcatVisualizer makes use of the new affordances in meshcat.
With meshcat, one can set the "opacity" or "modulated_opacity" property.
As with setting the "color" property, meshcat will update all of the
materials in a tree. Meshcat and meldis now exploit this to reduce the
number of meshcat calls (making a single API invocation for the entires
path tree that the instance owns).