Fix RenderEngineVtk upside down glTF textures (#20089)
It has been shown that vtkAssembly leads to textures being rendered
upside down. See
https://discourse.vtk.org/t/vtkgltfimporter-loads-textures-upside-down/12113
The fix we're applying is to get in front of that and flip the textures
ourselves when importing the glTF.

This PR further modifies rainbow_box.gltf so that its texture is no
longer vertically symmetric (therefore the GltfSupport test now tests
whether the texture is flipped or not).

It also provides a novel glTF file that contains all texture types and
an additional test to confirm that all of the textures are handled
appropriately. It uses a texture image that was generated *by* the test,
but on a local machine. Its correctness relies on a human looking at the
image and confirming that all of the textures are correctly aligned.