RenderEngineGltfClient changes how input glTFs are handled
There are two significant behavioral changes -- that *could* affect a user.
However, given the newness of the underlying code and the fact that this
touches distant corners of glTF-dom, the odds that anyone *actually* gets
affected is small.

Scene merging. Previously, all input scenes were included and
scenes with the same name were merged. Now only default scenes are merged,
regardless of names.

Extras and extensions are properly handled.
  - We now include two root arrays that were previously skipped:
    "extensionsUsed" and "extensionsRequired". These are now merged
    (without duplicates). The previous omission of "extensionsUsed" and
    "extensionsRequired" wasn't *obviously* a defect before because some
    glTF parsers are more forgiving than others in terms of *applying*
    extensions without having declared them as "used" or "required". This
    simply shores up the hole to make the final glTF as compliant as the
    input glTFs.
  - When it is necessary to merge extras or extensions, if there is
    ambiguity, we throw an exception with enough information to help the
    users address the conflict in their data. If we had been previously
    consuming glTF files that met the error conditions, they will now
    start throwing. This is, however, highly unlikely.