[symbolic] Give the default Variable a name (#20164)
Remove various restrictions on using the default Variable in an
Expression, Environment, Formula, etc.

This will be useful for univariate monomials and polynomials, where
the variable name doesn't matter but shouldn't be empty.

This is also probably a tiny speed-up to not hit the heap during the
default constructor (to allocate an empty reference-counted string).

Document that a moved-from Variable is equal to the default value.
This will make it easier to write move constructors for other classes
that contain variables but must maintain specific invariants.