RenderMaterial offers more nuanced and complete warnings (#20236)
The fallback materials would use the ('phong', 'diffuse_map') or matching
foo.png images as diffuse texture map, even if the underlying mesh could
not support it. This is because they were given no information about
whether a texture could be applied or not.

Now, it is provided information as to the support of texture coordinates
and will *also* omit a texture if the underlying geometry lacks sufficient
texture coordinate support.

Note: This leads to a very subtle, but non-destructive behavior change. If
a diffuse map was specified for a capsule in RenderEngineVtk, the image will
not appear (as it did before), but now there's a warning emitted.