Add DiscreteContactData (#20209)
We have three types of contact: point, hydroelastic, and deformable
that share similar contact data like  DiscreteContactPair and
ContactPairKinematics. Currently, we store them all in a std::vector
with an ordering convention that is inconsistent across various classes
that produce and consume these data.

We introduce DiscreteContactData to help enforce ordering conventions
across the board and ensure the correct contact data is retrieved.