[render] Parsing an .obj file can produce multiple RenderMesh instances (#20285)
* Parsing an .obj file can produce multiple RenderMesh instances (#20250)

Obj files have several mechanisms that would conceptually lead to creating
disjoint "meshes" (i.e., putting them in different groups, different
objects, or assigning different materials). For the purpose of rendering,
the only important distinction is material. Because we render the whole
mesh as a rigid object, we just need to partition based on the material
assigned to a triangle. LoadRenderMeshesFromObj can now return multiple
RenderMeshes.

We've introduced a shim for the old API (LoadRenderMeshFromObj()). This
allows us to update RenderEngineVtk and RenderEngineGl in a follow up PR.
The shim is *mostly* the same as the old. It has two differences which are
acceptable for a short horizon.

  - The debug message no longer prints out the material names.
  - The resultant RenderMesh can easily have an increased number of vertex
    data (as shown in the legacy test). This increase is representative of
    how things will work for the foreseeable future.