RgbdSensor now passes the correct camera pose
QueryObject documents that it expected the pose of the camera *body*
relative to a parent frame. RgbdSensor was passing the pose of the
camera sensor relative to the parent frame. There can be a non-identity
transform between the camera body and the camera sensor. The tests used
the identity transform so this mistake slipped through the cracks.