[math] Clarify implementation of fast pose composition functions (#20289)
Add more overview comments, adjusting all four functions to use a more
uniform explanatory style. In particular, introduce a vector lane
notation (explaining our assumption of low-to-high lane order) and the
idea of a "don't care" lane. Purge all comments that explain what
results are being computed in the don't care lanes. They were a lot of
clutter with no upside.

Rename the element names of the outputs (R_AC and X_AC) avoid the "ran
out of letters" problem and ensuing confusion with primes and multi-
character names. (We actually do have enough letters, if we sacrifice
one-letter names for the "<XYZ_> - <xyz_>" intermediate term in XinvX.)

For the broadcast (aka "four") coefficients, load the individual
doubles from memory instead of trying to broadcast one lane from a
prior 4-lane load.  Broadcasting from one lane of a vector (i.e.,
lane-crossing) is slow; Clang ends up respelling it to use a direct
memory access anyway, but GCC doesn't always find that peephole. We
might as well use the simpler spelling in the source code, especially
since it's what we want to happen, and gets GCC on board with the
faster instruction choices.

Reorder the code to do all loading first and all storing last. This is
better for readability and is what the compiler does anyway (better
for pipelining).

Reorder the code to keep each conceptual group (of dot products) in
one place, for clarity. The compiler reorders the machine code to be
optimally interleaved.

Reorder the XX and XinvX to start calculating the new translation
vector <stu_> first. This output has the deepest pipeline in terms of
flops, so it's important to get it started right away. Clang does this
automatically, but GCC chooses worse instruction sequences unless we
help it along.

According to llvm-mca with -cpu=skylake as the processor model, the
total cycles for these functions change to the following:

         Using Using
         GCC   Clang
         11.4  14.0
R@R
  master 34    28
  branch 29    28
R⁻¹@R
  master 34    30
  branch 30    30
X@X
  master 34    32
  branch 33    32
X⁻¹@X
  master 38    33
  branch 34    33

Notice how Clang is still the same, but GCC works better now.