[render] RenderEngineVtk supports multiple meshes from OBJ (#20235)
* RenderEngineVtk supports multiple meshes from OBJ

RenderEngineVtk uses the new LoadRenderMeshesFromObj() functionality. The
distinct meshes get mapped to the same prop/part infrastructure as the
glTF.

The description of obj support in the file format module is updated to
reflect the current capability.

Finally, it kills the transitory stub function introduced while
transition RenderEngineGl and Vtk to multi-part OBJs.