Bug fix to actuation input ordering.
Bug fix: Before this fix, it was assumed that joint actuators were added to a plant one model instance at a time.
This is true when we add model using our parsers. In this case the resulting actuation input for the full model is
ordered by model instance. This is no longer true if, for instance, a users adds actuators programatically in an
arbitrary order. This could lead to an actuation input vector that is no longer ordered by model instance.
This PR no longer assumes this order, but rather that the full actuation input is ordered by joint actuator index.
Unit tests are added to verify this assumption.

Co-authored-by: Jeremy Nimmer <jeremy.nimmer@tri.global>