[snopt] Nerf gfortran argument-mismatch errors into warnings
SNOPT 7.6 has broken code that gfortran previously accepted without
complaint but is now an error as of gfortran 10 or newer. Since the
code apparently works anyway, nerf the error down to a warning so that
users can still build with SNOPT 7.6.

(Drake does not officially supported building with 7.6, but since the
code for 7.4 is so much more difficult to obtain than 7.6 we try to
accommodate 7.6 on a best-effort basis.)