Fix mujoco parser joint kinematics. (#20414)
Resolves #20367.
Also adds support for joint defaults.

Together these support new tests confirming that the acrobot and
cartpole models from dm_control have the expected kinematics.

The MuJoCo XML documentation specifies the position of a joint as
being "specified in the frame of the parent body". But this apparently
needs to be interpretted as the current body, not the parent of the
current body.