Correct, test, and document FlattenModelDirectives
FlattenModelDirectives did not handle scoped names, which made it difficult to use in scenarios with multiple instances of the same model.

It was also undocumented and untested.

This change adds correct namespace handling; it also adds documentation and a minimal test (minimal because any naming-related errors are caught during the assembly of the MultibodyTree, and because the scoped test fixtures were designed to exercise naming).

 * Fixes #20425.
 * Removes misleading comments (including a mention of RigidBodyTree!).
 * Documents FlattenModelDirectives.
 * Documents the special-case naming semantics in add_frame directives.