[vtk_internal] Throw exceptions rather than abort()
The vtkXOpenGLRenderWindow.cxx initialization code aborts when a
DISPLAY cannot be found.  We are transforming this into an exception
because in a headless jupyter kernel (e.g., deepnote), the abort()
call makes the kernel restart without any helpful information to the
user as to _why_ it restarted.  With exceptions, the kernel will
continue to run but provide the exception message instead.